4565: Fixed issues following refactor
