+++ assets/admin/app.jsx
@@ -7,7 +7,7 @@ import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import localStorageKeys from "./components/util/local-storage-keys";
-import RestrictedRoute from "./restricted-route";
+import RestrictedRoute from "./components/restricted-route.jsx";
 import Topbar from "./components/navigation/topbar/top-bar";
 import SideBar from "./components/navigation/sidebar/sidebar";
 import ScreenList from "./components/screen/screen-list";
@@ -32,14 +32,14 @@ import UserContext from "./context/user-context";
 import ListContext from "./context/list-context";
 import SharedPlaylists from "./components/playlist/shared-playlists";
 import Logout from "./components/user/logout";
-import AuthHandler from "./auth-handler";
+import AuthHandler from "./components/auth-handler.jsx";
 import LoadingComponent from "./components/util/loading-component/loading-component";
 import ModalProvider from "./context/modal-context/modal-provider";
 import UsersList from "./components/users/users-list";
 import ActivationCodeList from "./components/activation-code/activation-code-list";
 import ActivationCodeCreate from "./components/activation-code/activation-code-create";
 import ActivationCodeActivate from "./components/activation-code/activation-code-activate";
-import AdminConfigLoader from "./admin-config-loader.js";
+import AdminConfigLoader from "./components/util/admin-config-loader.js";
 import "react-toastify/dist/ReactToastify.css";
 import "./app.scss";
 import FeedSourcesList from "./components/feed-sources/feed-sources-list";
