+++ assets/admin/components/slide/preview/slide-preview.jsx
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from "react";
 import { Button } from "react-bootstrap";
 import PropTypes from "prop-types";
 import { useTranslation } from "react-i18next";
-import ErrorBoundary from "../../../error-boundary";
+import ErrorBoundary from "../../error-boundary.jsx";
 import "./slide-preview.scss";
 import renderSlide from "../../../../template/render-slide.jsx";
