+++ assets/client/service/token-service.js
@@ -1,6 +1,6 @@
 import logger from "../logger/logger";
 import appStorage from "../util/app-storage";
-import ClientConfigLoader from "../client-config-loader.js";
+import ClientConfigLoader from "../util/client-config-loader.js";
 import defaults from "../util/defaults";
 import statusService from "./status-service";
 import constants from "../util/constants";
