+++ assets/template/screens.js
@@ -1,12 +1,12 @@
-import twoBoxes from "../screen-layouts/two-boxes.json";
-import threeBoxes from "../screen-layouts/three-boxes.json";
-import threeBoxesHorizontal from "../screen-layouts/three-boxes-horizontal.json";
-import twoBoxesVertical from "../screen-layouts/two-boxes-vertical.json";
-import touchTemplate from "../screen-layouts/touch-template.json";
-import sixAreas from "../screen-layouts/six-areas.json";
-import fullScreen from "../screen-layouts/full-screen.json";
-import fourAreas from "../screen-layouts/four-areas.json";
-import twoBoxesVerticalReversed from "../screen-layouts/two-boxes-vertical-reversed.json";
+import twoBoxes from "../shared/screen-layouts/two-boxes.json";
+import threeBoxes from "../shared/screen-layouts/three-boxes.json";
+import threeBoxesHorizontal from "../shared/screen-layouts/three-boxes-horizontal.json";
+import twoBoxesVertical from "../shared/screen-layouts/two-boxes-vertical.json";
+import touchTemplate from "../shared/screen-layouts/touch-template.json";
+import sixAreas from "../shared/screen-layouts/six-areas.json";
+import fullScreen from "../shared/screen-layouts/full-screen.json";
+import fourAreas from "../shared/screen-layouts/four-areas.json";
+import twoBoxesVerticalReversed from "../shared/screen-layouts/two-boxes-vertical-reversed.json";
 
 const screens = [
   {
