--- src/Template/TemplateInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Template;
-
-interface TemplateInterface
-{
-
-}
